requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.7
PyJWT==2.8.0

